+++ lib/pages/settings_page.dart
@@ -3,6 +3,9 @@ import 'package:dimensions_theme/dimensions_theme.dart';
 import 'package:in_app_review/in_app_review.dart';
 import 'package:share_plus/share_plus.dart';
 
+import 'package:taskzoo/misc/theme_notifier.dart';
+import 'package:taskzoo/widgets/settings/settings_option_widgets.dart';
+
 class SettingsPage extends StatefulWidget {
   const SettingsPage({Key? key}) : super(key: key);
 
@@ -175,110 +178,3 @@ class _SettingsPageState extends State<SettingsPage> {
   }
 }
 
-class SettingsOptionWithIcon extends StatefulWidget {
-  final IconData leftIcon;
-  final String optionText;
-  final IconData rightActionIcon;
-  final void Function() onActionTap;
-
-  SettingsOptionWithIcon({
-    required this.leftIcon,
-    required this.optionText,
-    required this.rightActionIcon,
-    required this.onActionTap,
-  });
-
-  @override
-  _SettingsOptionWithIconState createState() => _SettingsOptionWithIconState();
-}
-
-class _SettingsOptionWithIconState extends State<SettingsOptionWithIcon> {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.all(Dimensions.of(context).insets.small),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Row(
-            children: [
-              Icon(widget.leftIcon),
-              SizedBox(
-                width: Dimensions.of(context).insets.medium,
-              ),
-              Text(widget.optionText, style: TextStyle(fontSize: 16))
-            ],
-          ),
-          GestureDetector(
-            child: Icon(widget.rightActionIcon),
-            onTap: widget.onActionTap,
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class SettingsOptionWithToggle extends StatefulWidget {
-  final IconData leftIcon;
-  final String optionText;
-  final bool initialValue;
-  final ValueChanged<bool> onToggleChanged;
-
-  SettingsOptionWithToggle({
-    required this.leftIcon,
-    required this.optionText,
-    required this.initialValue,
-    required this.onToggleChanged,
-  });
-
-  @override
-  _SettingsOptionWithToggleState createState() =>
-      _SettingsOptionWithToggleState();
-}
-
-class _SettingsOptionWithToggleState extends State<SettingsOptionWithToggle> {
-  late bool _currentValue;
-
-  @override
-  void initState() {
-    super.initState();
-    _currentValue = widget.initialValue;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.all(Dimensions.of(context).insets.small),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Row(
-            children: [
-              Icon(widget.leftIcon),
-              SizedBox(
-                width: Dimensions.of(context).insets.medium,
-              ),
-              Text(widget.optionText, style: TextStyle(fontSize: 16))
-            ],
-          ),
-          Container(
-            constraints: BoxConstraints(
-                maxWidth: Theme.of(context).iconTheme.size! * 2,
-                maxHeight: Theme.of(context).iconTheme.size!),
-            child: Switch(
-              activeColor: Colors.black,
-              value: _currentValue,
-              onChanged: (bool value) {
-                setState(() {
-                  _currentValue = value;
-                });
-                widget.onToggleChanged(value);
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
